Fix some bugs in POS and Admin

+++ database/migrations/2021_11_24_092154_create_sales_reports_table.php
@@ -17,11 +17,11 @@ class CreateSalesReportsTable extends Migration
             $table->id('sales_report_id');
             $table->unsignedBigInteger('emp_id');
             $table->string('sales_invoice_no');
-            $table->decimal('total_price', 8, 2)->default(0.00);
-            $table->integer('cash')->nullable();
-            $table->integer('change')->nullable();
-            $table->integer('vatable_sale')->default(0.00)->comment('total price / 1.12')->nullable();
-            $table->integer('vat_amount')->default(0.00)->comment('vatable sale x 12%')->nullable();
+            $table->decimal('total_price', 15, 2)->default(0.00);
+            $table->decimal('cash',15,2)->nullable();
+            $table->decimal('change',15,2)->nullable();
+            $table->decimal('vatable_sale',15,2)->default(0.00)->comment('total price / 1.12')->nullable();
+            $table->decimal('vat_amount',15,2)->default(0.00)->comment('vatable sale x 12%')->nullable();
             $table->integer('status')->default(0)->comment('[0]-Pending [1]-Completed');
             $table->timestamps();
             $table->foreign('emp_id')
